% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwOmitNA.R
\name{dtwOmitNA}
\alias{dtwOmitNA}
\title{DTW Omit NA}
\usage{
dtwOmitNA(x, y)
}
\arguments{
\item{x}{first time sequence to compare}

\item{y}{second time sequence to compare}
}
\value{
returns the dynamic time warp distance between the two series
}
\description{
DTW Omit NA
}
