% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSyllables.R
\name{findSyllables}
\alias{findSyllables}
\title{Find Syllables
Find "syllables" (discrete acoustic events) in animal sounds}
\usage{
findSyllables(
  files,
  low = 1000,
  high = 22000,
  plotSyllables = FALSE,
  samplingRate = NULL,
  shortestPause = 25,
  shortestSyl = 10,
  plot = FALSE,
  sylThres = 0.5,
  windowLength = 20,
  overlap = 80
)
}
\arguments{
\item{files}{A vector of wave files}

\item{low}{lower frequency for filter}

\item{high}{lower frequency for filter}

\item{plotSyllables}{logical, should the spectrogram with the syllables be plotted?}

\item{samplingRate}{Sampling rate}

\item{shortestPause}{Smallest separation between syllables}

\item{shortestSyl}{Shortest event that cam be considered a syllables}

\item{plot}{should plot syllables according to soundgen?}

\item{sylThres}{percentage of mean amplitude envelope for audio event. Syllables are only detected when above this threshold}

\item{windowLength}{window size for audio events}

\item{overlap}{overlap between windows}
}
\value{
Dataframe with following collumns:
"sound.files" : original file from wich syllables were extracted,
"selec" : syllable id, unique inside the sound.file, but not between files,
"start" : syllable start,
"end" : syllable end
}
\description{
Find Syllables
Find "syllables" (discrete acoustic events) in animal sounds
}
\examples{
\dontrun{
songs = findSyllables(plotSyllables  = F,dir(here::here("inst/extdata"), full.names = TRUE,include.dirs = T), samplingRate = 44100)
 }

}
