% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSyllables.R
\name{plotSyllables}
\alias{plotSyllables}
\title{plotSyllables}
\usage{
plotSyllables(syls, low = 1000, high = 22000, labels = F)
}
\arguments{
\item{syls}{songs}

\item{low}{lowest frequency to plot}

\item{high}{highest frequency to plot}

\item{labels}{cluster labels}
}
\value{
returns nothing
}
\description{
plot spectrograms with syllables identified by findSyllables()
}
