% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveCutter.R
\name{waveCutter}
\alias{waveCutter}
\title{waveCutter}
\usage{
waveCutter(
  X,
  padding = 0.03,
  units = "seconds",
  low = 1000,
  high = 22000,
  write.wavs = TRUE,
  write.specs = FALSE
)
}
\arguments{
\item{X}{Vector of filenames}

\item{padding}{Time to be added before and after each selection, is seconds or milliseconds}

\item{units}{time unit for padding, either "seconds" (default) or "ms"}

\item{low}{in Hz, lower frequency limit for filter, optional}

\item{high}{in Hz, higher frequency limit for filter, optional}

\item{write.wavs}{logical, should the wav files for each selection be saved?}

\item{write.specs}{logical, should the spectrograms files for each selection be saved?}
}
\value{
Returns a dataframe with six columns: X = original filenames, "waves" = .wav filenames, "spectrograms" = spectrogram filenames
}
\description{
Takes a dataframe with event starts and ends (see output from findSyllables), and produces .wav files for each of these selections, along with a corresponding spectrogram
}
